Add NoItemsView

+++ TodoList/Veiws/ListView.swift
@@ -15,19 +15,25 @@ struct ListView: View {
     var body: some View {
         
         VStack {
-            List {
-                ForEach(listViewModel.items) { item in
-                    ListRowView(item: item)
-                        .onTapGesture {
-                            withAnimation(.linear) {
-                                listViewModel.updateItem(item: item)
-                            }
+            ZStack {
+                if listViewModel.items.isEmpty {
+                    Text("No items")
+                } else {
+                    List {
+                        ForEach(listViewModel.items) { item in
+                            ListRowView(item: item)
+                                .onTapGesture {
+                                    withAnimation(.linear) {
+                                        listViewModel.updateItem(item: item)
+                                    }
+                                }
                         }
+                        .onDelete(perform: listViewModel.deleteItem)
+                        .onMove(perform: listViewModel.moveItem)
+                    }
+                    .listStyle(PlainListStyle())
                 }
-                .onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
             }
-            .listStyle(PlainListStyle())
             .navigationTitle("Todo List 🖋️")
             .navigationBarItems(
                 leading: EditButton(),
